config handling errors

+++ src/controllers/character.js
@@ -3,6 +3,9 @@ const { Character, Movie } = require('../../database/models')
 const { Sequelize } = require('sequelize')
 const Op = Sequelize.Op
 
+const catchAsync = require('../utils/catchAsync')
+const ExpressError = require('../utils/ExpressError')
+
 exports.getAllCharacters = async (req, res) => {
     const { name, age, weight, movies } = req.query
     if (!name && !age && !weight && !movies) {
@@ -92,61 +95,42 @@ exports.getAllCharacters = async (req, res) => {
     }
 }
 
-exports.getOneCharacter = async (req, res) => {
-    try {
-        const { id } = req.params
-        const oneCharacter = await Character.findByPk(id)
-        return res.json(oneCharacter)
-    } catch (error) {
-        console.log(error)
-        res.status(500).json(error)
-    }
-}
+exports.getOneCharacter = catchAsync(async (req, res) => {
+    const { id } = req.params
+    if(!id) throw new ExpressError('No se ha encontrado el personaje', 400)
+    const oneCharacter = await Character.findByPk(id)
+    return res.json(oneCharacter)   
+})
 
-exports.createOneCharacter = async (req, res) => {
-    try {
-        const { image, name, age, weight, story } = req.body
-        const newCharacter = await Character.create({ image, name, age, weight, story})
-        return res.json(newCharacter)
-    } catch (error) {
-        console.log(error)
-        res.status(500).json(error)
-    } 
-}
+exports.createOneCharacter = catchAsync(async (req, res) => {
+    const { image, name, age, weight, story } = req.body
+    if(!image || !name || !age || !weight || !story) throw new ExpressError('Existen campos sin completar', 400)
+    const newCharacter = await Character.create({ image, name, age, weight, story})
+    return res.json(newCharacter)
+    
+})
 
-exports.updateOneCharacter = async (req, res) => {
-    try {
+exports.updateOneCharacter = catchAsync(async (req, res) => {
         const { id } = req.params
         const { image, name, age, weight, story } = req.body
+        if(!id) throw new ExpressError('No se ha encontrado el personaje', 400)
+        if(!image || !name || !age || !weight || !story) throw new ExpressError('Existen campos sin completar', 400)
         const editCharacter = await Character.update({ image, name, age, weight, story }, { where: { id }})
-        return res.json(editCharacter)
-    } catch (error) {
-        console.log(error)
-        res.status(500).json(error)
-    }
-}
+        return res.json(editCharacter)  
+})
 
-exports.deleteOneCharacter = async(req, res) => {
-    try {
-        const { id } = req.params
-        const deleteCharacter = await Character.destroy({
-            where: { id }
-        })
-        return res.json(deleteCharacter)
-    } catch (error) {
-        console.log(error)
-        res.status(500).json(error)
-    }
-}
+exports.deleteOneCharacter = catchAsync(async(req, res) => {
+    const { id } = req.params
+    if(!id) throw new ExpressError('No se ha encontrado el personaje', 400)
+    const deleteCharacter = await Character.destroy({
+        where: { id }
+    })
+    return res.json(deleteCharacter)
+})
 
-exports.deleteAllCharacter = async(req, res) => {
-    try {
+exports.deleteAllCharacter = catchAsync(async(req, res) => {
         const deleteAllCharacters = await Character.destroy({
             truncate : { cascade: true}
         })
-        return res.json(deleteAllCharacters)
-    } catch (error) {
-        console.log(error)
-        res.status(500).json(error)
-    }
-}
+        return res.json(deleteAllCharacters)  
+})
